Work-AI-Functions/requirements.txt
azure-devops
msrest
jira
GitPython
langchain
langgraph
llamaindex
python-dotenv